Fix errors when heaters aren't reported by the firmware
actual and target may be None if the firmware doesn't know a heater.
So far the plugin didn't honor these situations, leading to issue with
automatic shutoff e.g. in case of a printer without a heated bed or
in the future without a heated chamber report in the response to M105.

This patch makes things a bit more resilient.

See foosel/OctoPrint#3112